+++ packages/queue/src/queue/QueueManager.ts
@@ -14,8 +14,8 @@ export class QueueManager<
   private queueNames: EnumValue[];
 
   constructor(
-    private serviceNameAsPrefix: () => string,
-    private redisConfig: () => RedisOptions,
+    private serviceNameAsPrefix: string | (() => string),
+    private redisConfig: RedisOptions | (() => RedisOptions),
     private getLogger: () => ILogger,
     private getContext: () => SharedContext,
     private queueCreator: (
@@ -31,8 +31,12 @@ export class QueueManager<
   connectAll(): Promise<void> {
     const connManager = ConnectionManager.getInstance();
     connManager.setConfiguration(
-      this.redisConfig(),
-      this.serviceNameAsPrefix(),
+      typeof this.redisConfig === "function"
+        ? this.redisConfig()
+        : this.redisConfig,
+      typeof this.serviceNameAsPrefix === "function"
+        ? this.serviceNameAsPrefix()
+        : this.serviceNameAsPrefix,
       this.getLogger
     );